db_config.py
import mysql.connector
import os
from dotenv import load_dotenv

load_dotenv()

def get_connection():
    """
    Membuat koneksi ke database MySQL menggunakan environment variables.
    """
    try:
        conn = mysql.connector.connect(
            host=os.getenv('DB_HOST'),
            user=os.getenv('DB_USER'),
            password=os.getenv('DB_PASSWORD'),
            database=os.getenv('DB_NAME')
        )
        return conn
    except mysql.connector.Error as e:
        print(f"Error connecting to MySQL database: {e}")

        return None