requirements.txt
flask
mysql-connector-python
numpy
scikit-fuzzy
scipy
flask-basicauth
Flask-WTF
python-dotenv